MUU-674: merge/muuntaja common submodules and fixes (#319)
* Unified naming schema using event

* Data handling and transformed update to muuntaja/merge
- common uses data module with export/import functions
- transformed wont be directly modified anymore
- bunch of get/update functions
- in muuntaja setting options from server uses tags now instead of objects
   - these seem redundant

* Common url module for muuntaja/merge

* Common modules added, some functions moved to modules:
- created some handler files that are imported as modules
  - modules have init functions preset in case of need and to make sure ide import clean wont clear module imports with only window functions
- moved some functions to modules from common
- data file rename to xHandler format
- all modules use init function, even when just placeholder

* Fix: data import filename update

* Transformation module, init logs, some cleanup:
- transformation related functions moved to own handler with required imports
- added console logs for module inits just in case
- ran format documents for scritps

* Further submodules for transformation common module:
- ui functions separated to edit based and common modules
- field click functions moved to ui edit module
- show transformed now available via global window object
- cleaned up common script (just about only inits and module export remaining)

* Transformation datamodule config change:
- any and all required client behaviour configs for client passed from merge/muuntaja are within object
- added config for keeping base record when using new instance

* Edit handler additional feature, on new instance update:
- edit handler has functionality to turn off edit mode, it updates the data itself and can now call to update edit mode visual
- edit mode visuals are now available to be updated outside of click event
- when using new operation the edit mode is from data and visual turned off
- when using new operation if the config says so baseinput is also cleared

* Doocument formating and datamodule reference fixes

* Copy link update:
- url already has needed info, we just need to make sure its valid
- since now "url update" also sanitizes url call it to make sure url is valid and up to date, then use existing url, no need to add profile and type separately
